+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.5.9
+
+- remove `repeatCount` prop from `@TimeSwitcher`
+- support of multiple `@TimeSwitcher` in Matlab - draft
+- 
+
 ## 0.5.8
 
 - add spaceFilter for Mrgsolve, Julia